% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg_covariance.R
\name{bandwidth}
\alias{bandwidth}
\title{Bandwidth}
\usage{
bandwidth(n, alpha, tau = 0.05, type = "Hall-Sheather")
}
\arguments{
\item{n}{sample size}

\item{alpha}{quantile index}

\item{tau}{alpha level of the test}

\item{type}{Bofinger, Chamberlain or Hall-Sheather}
}
\description{
Bandwidth for the estimation of the density quantile function
}
\references{
Bofinger (1975), Chamberlain (1994), Hall and Sheather(1988)
}
