% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{bread.esreg}
\alias{bread.esreg}
\title{Bread in the sandwich}
\usage{
\method{bread}{esreg}(x, sparsity = "nid",
  bandwidth_estimator = "Hall-Sheather", misspec = TRUE, ...)
}
\arguments{
\item{x}{An \link{esreg} object}

\item{sparsity}{The estimator to be used for the sparsity in \eqn{\Lambda} (the bread in the sandwich), see \link{density_quantile_function}
\itemize{
  \item iid - Piecewise linear interpolation of the distribution
  \item nid - Hendricks and Koenker sandwich
  \item indicator - TBA
}}

\item{bandwidth_estimator}{The bandwidth estimator to be used for the iid and nid sparsity estimator, see \link{density_quantile_function}
\itemize{
  \item Bofinger
  \item Chamberlain
  \item Hall-Sheather
}}

\item{misspec}{if TRUE, the estimator accounts for potential misspecification in the model, see TBA}

\item{...}{Further arguments (does not apply here)}
}
\description{
The bread in the sandwich, i.e. the inverse of \eqn{\Lambda}, for more details see \link{lambda_matrix}.
This function is added for compatibility with the sandwich package.
}
