% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_quantities.R
\name{cdf_at_quantile}
\alias{cdf_at_quantile}
\title{Cumulative Density Function at Quantile}
\usage{
cdf_at_quantile(y, x, q)
}
\arguments{
\item{y}{Vector of dependent data}

\item{x}{Matrix of covariates including the intercept}

\item{q}{Vector of quantile predictions}
}
\description{
Returns the cumulative density function evaluated at quantile predictions.
For a correctly specified model this should yield a value close to the quantile level.
}
\keyword{internal}
