% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_quantities.R
\name{conditional_mean_sigma}
\alias{conditional_mean_sigma}
\title{Conditional Mean and Sigma}
\usage{
conditional_mean_sigma(y, x)
}
\arguments{
\item{y}{Vector of dependent data}

\item{x}{Matrix of covariates including the intercept}
}
\description{
Estimate the conditional mean and sigma of the dependent data conditional on covariates x
}
\keyword{internal}
