% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg_covariance.R
\name{conditional_truncated_variance}
\alias{conditional_truncated_variance}
\title{Conditional truncated variance}
\usage{
conditional_truncated_variance(y, x, approach)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Covariates}

\item{approach}{scl_N or scl_t}
}
\description{
Estimate the conditional truncated variance
}
