% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_quantities.R
\name{density_quantile_function}
\alias{density_quantile_function}
\title{Density Quantile Function}
\usage{
density_quantile_function(y, x, u, alpha, sparsity, bandwidth_type)
}
\arguments{
\item{y}{Vector of dependent data}

\item{x}{Matrix of covariates}

\item{u}{Quantile residuals}

\item{alpha}{Probability level}

\item{sparsity}{iid or ind}

\item{bandwidth_type}{Bofinger, Chamberlain or Hall-Sheather}
}
\description{
Estimate the density quantile function
}
\references{
Bofinger (1975), Chamberlain (1994), Hall and Sheather(1988)
}
