% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{esr_loss}
\alias{esr_loss}
\title{Joint (VaR, ES) loss function}
\usage{
esr_loss(r, q, e, alpha, g1 = 2L, g2 = 1L, return_mean = TRUE)
}
\arguments{
\item{r}{Vector of returns}

\item{q}{Vector of quantiles}

\item{e}{Vector of expected shortfalls}

\item{alpha}{Probability level}

\item{g1}{1, 2, see \link{G1_fun}}

\item{g2}{1, 2, 3, 4, 5, see \link{G2_curly_fun}, \link{G2_fun}}

\item{return_mean}{If TRUE returns the average tick loss, else the individual values}
}
\description{
Computes the joint (VaR, ES) loss
}
\references{
Fissler and Ziegel (2016)
}
