% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{esr_rho_lp}
\alias{esr_rho_lp}
\title{Joint (VaR, ES) loss for a linear predictor}
\usage{
esr_rho_lp(b, y, x, alpha, g1 = 2L, g2 = 1L, delta = 0)
}
\arguments{
\item{b}{Parameter vector}

\item{y}{Vector of dependent data}

\item{x}{Matrix of covariates. Note: intercept needs to be added manually}

\item{alpha}{Quantile of interest}

\item{g1}{1, 2 (see \link{G1_fun})}

\item{g2}{1, 2, 3, 4, 5 (see \link{G2_curly_fun}, \link{G2_fun})}

\item{delta}{Approximation of the indicator function (0 is the indicator function)}
}
\description{
Returns the loss at b
}
