% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg_covariance.R
\name{esreg_covariance}
\alias{esreg_covariance}
\title{Asymptotic covariance}
\usage{
esreg_covariance(fit, sparsity = "iid", cond_var = "ind",
  bandwidth_type = "Hall-Sheather")
}
\arguments{
\item{fit}{fit An object from calling esreg()}

\item{sparsity}{Sparsity estimator
\itemize{
   \item iid - Piecewise linear interpolation of the distribution
   \item nid - Hendricks and Koenker sandwich (two additional quantile regressions)
}}

\item{cond_var}{Conditional truncated variance estimator
\itemize{
   \item ind Variance over all negative residuals
   \item scl_N Scaling with the Normal distribution
   \item scl_t Scaling with the t-distribution
}}

\item{bandwidth_type}{Bofinger, Chamberlain or Hall-Sheather}
}
\description{
Asymptotic covariance
}
