% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{esreg_covariance_boot}
\alias{esreg_covariance_boot}
\title{Bootstrapped Covariance}
\usage{
esreg_covariance_boot(fit, B = 1000, bootstrap_method = "iid",
  block_length = NULL)
}
\arguments{
\item{fit}{esreg object}

\item{B}{Number of bootstrap iterations}

\item{bootstrap_method}{iid or stationary}

\item{block_length}{Average block length for the stationary bootstrap}
}
\description{
Estimate the variance-covariance matrix via bootstrapping
}
\references{
Politis & Romano (1994)
}
