% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{estfun.esreg}
\alias{estfun.esreg}
\title{Estimating function}
\usage{
estfun.esreg(x, ...)
}
\arguments{
\item{x}{An \link{esreg} object}

\item{...}{Further arguments (does not apply here)}
}
\description{
This function matches the estfun function of the sandwich package and
returns the estimating functions for the fitted model.
It can for instance be used for an OPG estimator of the sigma matrix.
For esreg, the dimension of the estimating functions is n x (kq + ke).
}
