% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G1_fun}
\alias{G1_fun}
\title{Specification Function}
\usage{
G1_fun(z, type)
}
\arguments{
\item{z}{Data}

\item{type}{Choice of the G1 function:
\itemize{
  \item 1: G1(z) = z
  \item 2: G1(z) = 0
}}
}
\description{
G1
}
\keyword{internal}
