% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G1_prime_fun}
\alias{G1_prime_fun}
\title{Specification Function}
\usage{
G1_prime_fun(z, type)
}
\arguments{
\item{z}{Data}

\item{type}{Choice of the G1_prime function:
\itemize{
  \item 1: G1_prime(z) = 1
  \item 2: G1_prime(z) = 0
}}
}
\description{
G1_prime
}
\keyword{internal}
