% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G2_curly_fun}
\alias{G2_curly_fun}
\title{Specification Function}
\usage{
G2_curly_fun(z, type)
}
\arguments{
\item{z}{Data}

\item{type}{Choice of the G2_curly function:
\itemize{
  \item 1: -log(-z), z < 0
  \item 2: -sqrt(-z), z < 0
  \item 3: -1/z, z < 0
  \item 4: log(1 + exp(z))
  \item 5: exp(z)
}}
}
\description{
G2_curly
}
\keyword{internal}
