% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G2_fun}
\alias{G2_fun}
\title{Specification Function}
\usage{
G2_fun(z, type)
}
\arguments{
\item{z}{Data}

\item{type}{Choice of the G2 function:
\itemize{
  \item 1: -1/z, z < 0
  \item 2: 0.5/sqrt(-z), z < 0
  \item 3: 1/z^2, z < 0
  \item 4: 1 / (1 + exp(-z))
  \item 5: exp(z)
}}
}
\description{
G2
}
\keyword{internal}
