% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G2_prime_fun}
\alias{G2_prime_fun}
\title{Specification Function}
\usage{
G2_prime_fun(z, type)
}
\arguments{
\item{z}{Data}

\item{type}{Choice of the G2_prime function:
\itemize{
  \item 1: 1/z^2, z < 0
  \item 2: 0.25 / (-z)^(3/2), z < 0
  \item 3: -2/z^3, z < 0
  \item 4: exp(z) / (1 + exp(z))^2
  \item 5: exp(z)
}}
}
\description{
G2_prime
}
\keyword{internal}
