% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{gpl}
\alias{gpl}
\title{Generalized piecewise linear loss function}
\usage{
gpl(r, q, alpha, g = function(x) x, return_mean = TRUE)
}
\arguments{
\item{r}{Vector of returns}

\item{q}{Vector of quantiles}

\item{alpha}{Probability level}

\item{g}{A nondecreasing function}

\item{return_mean}{If TRUE returns the average tick loss, else the individual values}
}
\description{
Equivalent to the tick / check loss when g is the identity function.
}
\references{
Gneiting (2011)
}
