% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{lambda_matrix}
\alias{lambda_matrix}
\title{Lambda Matrix}
\usage{
lambda_matrix(object, sparsity, bandwidth_estimator, misspec)
}
\arguments{
\item{object}{An esreg object}

\item{sparsity}{The estimator to be used for the sparsity in \eqn{\Lambda}, see \link{density_quantile_function}
\itemize{
  \item iid - Piecewise linear interpolation of the distribution
  \item nid - Hendricks and Koenker sandwich
}}

\item{bandwidth_estimator}{The bandwidth estimator to be used for the iid and nid sparsity estimator, see \link{density_quantile_function}
\itemize{
  \item Bofinger
  \item Chamberlain
  \item Hall-Sheather
}}

\item{misspec}{if TRUE, the estimator accounts for potential misspecification in the model}
}
\description{
Estimate the lambda matrix.
}
