% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{model.matrix.esreg}
\alias{model.matrix.esreg}
\title{Model Matrix}
\usage{
\method{model.matrix}{esreg}(object, ...)
}
\arguments{
\item{object}{An \link{esreg} object}

\item{...}{Further arguments (does not apply here)}
}
\description{
Returns a n x (kq + ke) matrix with covariates in order to match \link{estfun.esreg}
and its use in \link[sandwich]{meatHAC}.
}
