% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{sigma_matrix}
\alias{sigma_matrix}
\title{Sigma Matrix ("meat")}
\usage{
sigma_matrix(object, sigma_est)
}
\arguments{
\item{object}{An esreg object}

\item{sigma_est}{The estimator to be used for \eqn{\Sigma} (the meat in the sandwich)
\itemize{
  \item ind - Variance over all negative residuals (see \link{conditional_truncated_variance})
  \item scl_N - Scaling with the normal distribution (see \link{conditional_truncated_variance})
  \item scl_sp - Scaling with the kernel density function (see \link{conditional_truncated_variance})
  \item OPG - outer product of the estimating functions (see \link[sandwich]{meat})
  \item OPG-HAC - HAC version of the outer product of the estimating functions (see \link[sandwich]{meatHAC})
  }}
}
\description{
Estimate the sigma matrix (the meat in the sandwich).
}
