% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{sigma_matrix_scaling}
\alias{sigma_matrix_scaling}
\title{Estimate the sigma matrix (the meat in the sandwich) using the scaling estimator.}
\usage{
sigma_matrix_scaling(object, sigma_est, misspec)
}
\description{
Estimate the sigma matrix (the meat in the sandwich) using the scaling estimator.
}
\keyword{internal}
