% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{summary.esreg}
\alias{summary.esreg}
\title{esreg summary}
\usage{
\method{summary}{esreg}(object, ...)
}
\arguments{
\item{object}{An esreg object}

\item{...}{Accepts all parameters you can pass to \code{\link{vcov.esreg}}.}
}
\description{
Summarize details about the regression estimates.
}
