% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg_twostep.R
\name{summary.esreg_twostep}
\alias{summary.esreg_twostep}
\title{esreg_twostep summary}
\usage{
\method{summary}{esreg_twostep}(object, ...)
}
\arguments{
\item{object}{An esreg_twostep object}

\item{...}{Accepts all parameters you can pass to \code{\link{vcov.esreg_twostep}}.}
}
\description{
Summarize details about the regression estimates.
}
\keyword{internal}
