% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{vcov.esreg}
\alias{vcov.esreg}
\title{Covariance Estimation}
\usage{
\method{vcov}{esreg}(object, method = "asymptotic", ...)
}
\arguments{
\item{object}{An \link{esreg} object}

\item{method}{For asymptotic use \link{vcovA}, for boot use \link{vcovB}}

\item{...}{All possible values which can be passed to \link{vcovA} and \link{vcovB}}
}
\description{
Estimate the variance-covariance matrix of the joint (VaR, ES) estimator
}
