% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{vcovB}
\alias{vcovB}
\title{Bootstrap Covariance Estimation}
\usage{
vcovB(object, bootstrap_method = "iid", B = 1000)
}
\arguments{
\item{object}{An esreg object}

\item{bootstrap_method}{The bootstrap sampling scheme to be used
\itemize{
  \item iid - The iid bootstrap of Efron (1979)
}}

\item{B}{The number of bootstrap iterations}
}
\description{
Estimate the variance-covariance matrix of the joint (VaR, ES) estimator
using the bootstrap.
}
