% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{vcovMS}
\alias{vcovMS}
\title{Covariance Estimation under Misspecification}
\usage{
vcovMS(object, sparsity = "nid", bandwidth_type = "Hall-Sheather",
  cond_var = "nid")
}
\arguments{
\item{object}{An esreg object}
}
\description{
Estimate the variance-covariance matrix of the joint (VaR, ES) estimator
using TODO.
}
