% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_quantities.R
\name{conditional_truncated_variance}
\alias{conditional_truncated_variance}
\title{Conditional truncated variance}
\usage{
conditional_truncated_variance(y, x, approach)
}
\arguments{
\item{y}{Vector of dependent data}

\item{x}{Matrix of covariates}

\item{approach}{ind, scl_N or scl_t}
}
\description{
Estimate the conditional truncated variance of y given x when y is truncated at 0.
}
