% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_quantities.R
\name{conditional_truncated_variance}
\alias{conditional_truncated_variance}
\title{Conditional truncated variance}
\usage{
conditional_truncated_variance(y, x, approach)
}
\arguments{
\item{y}{vector of dependent data}

\item{x}{matrix of covariates}

\item{approach}{ind, scl_N or scl_t}

\item{u}{quantile residuals}
}
\description{
Estimate the conditional truncated variance
}
