% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_quantities.R
\name{conditional_truncated_variance}
\alias{conditional_truncated_variance}
\title{Conditional truncated variance}
\usage{
conditional_truncated_variance(y, x, approach, approximate = TRUE)
}
\arguments{
\item{y}{Vector of dependent data}

\item{x}{Matrix of covariates including the intercept}

\item{approach}{ind, scl_N, scl_t or scl_st}

\item{approximate}{Concerns the scl_st estimator: approximate the time consuming
integration via splines.}
}
\description{
Estimate the conditional truncated variance of y given x
and given y <= 0.
}
