% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_quantities.R
\name{conditional_truncated_variance}
\alias{conditional_truncated_variance}
\title{Conditional truncated variance}
\usage{
conditional_truncated_variance(y, x, approach)
}
\arguments{
\item{y}{Vector of dependent data}

\item{x}{Matrix of covariates including the intercept}

\item{approach}{ind, scl_N or scl_sp}
}
\description{
Estimate the variance of y given x and given y <= 0.
}
