% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{esr_rho_lp}
\alias{esr_rho_lp}
\title{Joint (VaR, ES) loss for a linear predictor}
\usage{
esr_rho_lp(b, y, xq, xe, alpha, g1 = 2L, g2 = 1L)
}
\arguments{
\item{b}{Parameter vector}

\item{y}{Vector of dependent data}

\item{xq}{Matrix of covariates for the quantile part}

\item{xe}{Matrix of covariates for the expected shortfall part}

\item{alpha}{Probability level}

\item{g1}{1, 2 (see \link{G1_fun})}

\item{g2}{1, 2, 3, 4, 5 (see \link{G2_curly_fun}, \link{G2_fun})}
}
\description{
Returns the loss for the parameter vector b
}
\keyword{internal}
