% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{esreg}
\alias{esreg}
\title{Joint (VaR, ES) Regression}
\usage{
esreg(formula, data, alpha, g1 = 2L, g2 = 1L, b0 = NULL, target = "rho",
  method = "random_restart", random_restart_ctrl = list(M = 10, N = 10000,
  sd = sqrt(0.1)), gensa_ctrl = list(max.call = 1e+06, max.time = 5, box =
  10))
}
\arguments{
\item{formula}{y ~ x1 + x2 + ...}

\item{data}{data.frame that stores y and x. Extracted from the enviroment if missing.}

\item{alpha}{Quantile index}

\item{g1}{1, 2 (see \link{G1_fun}, \link{G1_prime_fun})}

\item{g2}{1, 2, 3, 4, 5 (see \link{G2_curly_fun}, \link{G2_fun}, \link{G2_prime_fun})}

\item{b0}{Starting values; if NULL they are obtained from two additional quantile regressions}

\item{target}{The functions to be optimized: either the loss (rho) or the identification function (psi).}

\item{method}{one_shot, random_restart or gensa}

\item{random_restart_ctrl}{list: N number of random starting points; M optimize over the M best; sd std dev of the random component}

\item{gensa_ctrl}{Parameters to be passed to GenSA}
}
\description{
Joint (VaR, ES) Regression
}
