% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg_covariance.R
\name{esreg_covariance_boot}
\alias{esreg_covariance_boot}
\title{Bootstrapped Covariance}
\usage{
esreg_covariance_boot(fit, B = 1000, block_length = 1,
  method = "one_shot")
}
\arguments{
\item{fit}{esreg object}

\item{B}{bootstrap iterations}

\item{block_length}{Length of each block, 1 corresponds to iid bootstrap}

\item{method}{Same as for esreg(). Defaults to one_shot for speed.}
}
\description{
Estimate the variance-covariance matrix using the stationary block bootstrap
}
\references{
Politis & Romano (1994)
}
