% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{esreg_twostep}
\alias{esreg_twostep}
\title{Two Step (VaR, ES) Regression}
\usage{
esreg_twostep(formula, data, alpha)
}
\arguments{
\item{formula}{y ~ x1 + x2 + ...}

\item{data}{data.frame that stores y and x. Extracted from the enviroment if missing.}

\item{alpha}{Quantile index}
}
\description{
Estimates the expected shortfall in two steps.
}
\details{
This estimator is much faster than the
one-step estimator \link{esreg}. Its estimates are, however,
less precise and therefore \link{esreg} is geneally the preferred estimator.
}
