% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg_twostep.R
\name{esreg_twostep}
\alias{esreg_twostep}
\title{Two Step (VaR, ES) Regression}
\usage{
esreg_twostep(formula, data, alpha)
}
\arguments{
\item{formula}{Formula object, e.g.: y ~ x1 + x2 + ...}

\item{data}{data.frame that holds the variables. Can be missing.}

\item{alpha}{Quantile of interest}

\item{...}{additional arguments}
}
\description{
Estimates the expected shortfall in two steps. First a linear quantile regression,
then a weighted least squares regression. This estimator is much faster than the joint.
The estimates are, however, often less precise. For more information, see \code{\link{esreg}}.
}
\seealso{
\code{\link{vcov.esreg_twostep}} for the covariance estimation and
\code{\link{summary.esreg_twostep}} for a summary of the regression results
}
