% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{estfun.esreg}
\alias{estfun.esreg}
\title{Estimating function}
\usage{
\method{estfun}{esreg}(x, ...)
}
\arguments{
\item{x}{An \link{esreg} object}

\item{...}{Further arguments (does not apply here)}
}
\description{
This function matches the \link[sandwich]{estfun} function of the sandwich package and
returns the estimating functions for the fitted model.
For esreg, the dimension of the estimating functions is n x (kq + ke).
}
