% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{sigma_matrix}
\alias{sigma_matrix}
\title{Sigma Matrix}
\usage{
sigma_matrix(object, sigma_est, misspec)
}
\arguments{
\item{object}{An esreg object}

\item{sigma_est}{The estimator to be used for \eqn{\Sigma}, see \link{conditional_truncated_variance}
\itemize{
  \item ind - Variance over all negative residuals
  \item scl_N - Scaling with the normal distribution
  \item scl_sp - Scaling with the kernel density function
  }}

\item{misspec}{if TRUE, the estimator accounts for potential misspecification in the model}
}
\description{
Estimate the sigma matrix.
}
