% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{vcov.esreg}
\alias{vcov.esreg}
\title{Variance-Covariance Matrix}
\usage{
\method{vcov}{esreg}(object, boot = FALSE, ...)
}
\arguments{
\item{object}{esreg object}

\item{boot}{If TRUE, bootstrap the covariance; else use the asymptotic formulas}

\item{...}{other paramters passed to the low-level functions}
}
\description{
Variance-Covariance Matrix
}
