% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{vcov.esreg}
\alias{vcov.esreg}
\title{Covariance Estimation for esreg}
\usage{
\method{vcov}{esreg}(object, type = "asymptotic", sparsity = "iid",
  cond_var = "ind", bandwidth_type = "Hall-Sheather",
  bootstrap_method = "iid", B = 1000, ...)
}
\arguments{
\item{object}{An esreg object}

\item{type}{Either asymptotic or bootstrap.}

\item{sparsity}{The sparsity estimator,
see \link{density_quantile_function} for more details.
\itemize{
  \item iid - Piecewise linear interpolation of the distribution
  \item nid - Hendricks and Koenker sandwich
}}

\item{cond_var}{The conditional truncated variance estimator,
see \link{conditional_truncated_variance} for more details.
\itemize{
  \item ind - Variance over all negative residuals
  \item scl_N - Scaling with the normal distribution
  \item scl_sp - Scaling with the kernel density function
  }}

\item{bandwidth_type}{The bandwidth estimator
\itemize{
  \item Bofinger
  \item Chamberlain
  \item Hall-Sheather
}}

\item{bootstrap_method}{The bootstrap sampling scheme to be used
\itemize{
  \item iid - The iid bootstrap of Efron (1979)
  \item stationary - The stationary bootstrap of Politis & Romano (1994)
}}

\item{B}{The number of bootstrap iterations}

\item{...}{Further arguments (does not apply here)}

\item{block_length}{Average block length for the stationary bootstrap}
}
\description{
Estimate the variance-covariance matrix of the joint (VaR, ES) estimator either
using the asymptotic formulas or using the bootstrap.
}
