% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{vcov.esreg}
\alias{vcov.esreg}
\title{Covariance of the joint (VaR, ES) estimator}
\usage{
\method{vcov}{esreg}(object, sparsity = "iid", cond_var = "ind",
  bandwidth_type = "Hall-Sheather", bootstrap_method = NULL, B = 1000,
  block_length = NULL, ...)
}
\arguments{
\item{object}{An esreg object}

\item{sparsity}{Sparsity estimator
\itemize{
  \item [iid] - Piecewise linear interpolation of the distribution
  \item nid - Hendricks and Koenker sandwich
}}

\item{cond_var}{Conditional truncated variance estimator
\itemize{
  \item [ind] Variance over all negative residuals
  \item scl_N Scaling with the normal distribution
  \item scl_sp Scaling with the kernel density function
}}

\item{bandwidth_type}{Bofinger, Chamberlain or Hall-Sheather}

\item{bootstrap_method}{\itemize{
  \item [NULL] Use the asymptotic estimator
  \item iid bootstrap
  \item stationary bootstrap (Politis & Romano, 1994)
}}

\item{B}{Number of bootstrap iterations}

\item{block_length}{Average block length for the stationary bootstrap}

\item{...}{additional arguments}
}
\description{
Estimate the variance-covariance matrix of the joint (VaR, ES) estimator either using the asymptotic formulas or using the bootstrap.
}
