% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esreg.R
\name{vcovA}
\alias{vcovA}
\title{Asymptotic Covariance Estimation}
\usage{
vcovA(object, sigma_est = "OPG-HAC", sparsity = "nid",
  misspec = TRUE, bandwidth_estimator = "Hall-Sheather")
}
\arguments{
\item{object}{An esreg object}

\item{sigma_est}{The estimator to be used for \eqn{\Sigma} (the meat in the sandwich)
\itemize{
  \item ind - Variance over all negative residuals (see \link{conditional_truncated_variance})
  \item scl_N - Scaling with the normal distribution (see \link{conditional_truncated_variance})
  \item scl_sp - Scaling with the kernel density function (see \link{conditional_truncated_variance})
  \item OPG - outer product of the estimating functions (see \link[sandwich]{meat})
  \item OPG-HAC - HAC version of the outer product of the estimating functions (see \link[sandwich]{meatHAC})
  }}

\item{sparsity}{The estimator to be used for the sparsity in \eqn{\Lambda} (the bread in the sandwich), see \link{density_quantile_function}
\itemize{
  \item iid - Piecewise linear interpolation of the distribution
  \item nid - Hendricks and Koenker sandwich
  \item indicator - TBA
}}

\item{misspec}{if TRUE, the estimator accounts for potential misspecification in the model, see TBA}

\item{bandwidth_estimator}{The bandwidth estimator to be used for the iid and nid sparsity estimator, see \link{density_quantile_function}
\itemize{
  \item Bofinger
  \item Chamberlain
  \item Hall-Sheather
}}
}
\description{
Estimate the variance-covariance matrix of the joint (VaR, ES) estimator by the sandwich formula:
\deqn{\lambda^{-1} \Sigma \lambda^{-1}}
Several estimators are available for both matrices and the default options are selected to take into account
possible misspecifications in the underlying data.
}
